move Inventory services to de.metas.inventory
https://github.com/metasfresh/metasfresh/issues/2647
