+++ de.metas.business/src/main/java/de/metas/inventory/IInventoryDAO.java
@@ -1,4 +1,4 @@
-package org.adempiere.minventory.api;
+package de.metas.inventory;
 
 import java.util.List;
