+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/inventory/impl/HUInventoryBL.java
@@ -5,7 +5,6 @@ import java.util.Collection;
 import java.util.List;
 
 import org.adempiere.ad.trx.api.ITrx;
-import org.adempiere.minventory.api.IInventoryDAO;
 import org.adempiere.util.GuavaCollectors;
 import org.adempiere.util.Services;
 import org.compiere.model.I_M_InventoryLine;
@@ -18,6 +17,7 @@ import de.metas.handlingunits.IHUAssignmentDAO;
 import de.metas.handlingunits.inventory.IHUInventoryBL;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_M_Inventory;
+import de.metas.inventory.IInventoryDAO;
 
 /*
  * #%L
