+++ de.metas.swat/de.metas.swat.base/src/main/java/de/metas/invoicecandidate/spi/impl/M_Inventory_Handler.java
@@ -5,7 +5,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Properties;
 
-import org.adempiere.minventory.api.IInventoryDAO;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.util.Services;
 import org.compiere.model.I_M_Inventory;
@@ -13,6 +12,7 @@ import org.compiere.model.I_M_InventoryLine;
 
 import com.google.common.collect.ImmutableList;
 
+import de.metas.inventory.IInventoryDAO;
 import de.metas.invoicecandidate.api.IInvoiceCandidateHandlerBL;
 import de.metas.invoicecandidate.model.I_C_Invoice_Candidate;
 import de.metas.invoicecandidate.spi.AbstractInvoiceCandidateHandler;
